# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Papertrail API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ingest/index": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Index Status",
        "description": "Image counts and the batches the chosen indexing scheme would add.",
        "operationId": "index_status_api_ingest_index_get",
        "parameters": [
          {
            "name": "scheme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Confirm Index",
        "description": "Add the proposed batches to batches.json (only if the proposal is unchanged).",
        "operationId": "confirm_index_api_ingest_index_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIndexIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ingest/pages/toss": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Toss Page",
        "description": "Toss the document this page belongs to.",
        "operationId": "toss_page_api_ingest_pages_toss_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveGroupingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/pages/recover": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Recover Page",
        "description": "Undo a toss.",
        "operationId": "recover_page_api_ingest_pages_recover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveGroupingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/pages/rotate": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Rotate Page",
        "description": "Rotate a page's scan in place so it is upright.",
        "operationId": "rotate_page_api_ingest_pages_rotate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveGroupingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/ocr": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Ocr Status",
        "description": "Page counts for the chosen scope; ``provider`` defaults to the saved OCR model.",
        "operationId": "ocr_status_api_ingest_ocr_get",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "batch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch Id"
            }
          },
          {
            "name": "reprocess",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Reprocess"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Start Ocr",
        "description": "Start OCR on the planned pages as a background job.",
        "operationId": "start_ocr_api_ingest_ocr_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartOcrIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/parse": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Parse Status",
        "operationId": "parse_status_api_ingest_parse_get",
        "parameters": [
          {
            "name": "reprocess",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Reprocess"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Start Parse",
        "description": "Start extraction on the planned documents as a background job.",
        "operationId": "start_parse_api_ingest_parse_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartParseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmIndexIn": {
        "properties": {
          "scheme": {
            "type": "string",
            "title": "Scheme"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scheme",
          "token"
        ],
        "title": "ConfirmIndexIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexStatus": {
        "properties": {
          "blocker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker"
          },
          "schemes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Schemes"
          },
          "scheme": {
            "type": "string",
            "title": "Scheme"
          },
          "image_count": {
            "type": "integer",
            "title": "Image Count"
          },
          "indexed_count": {
            "type": "integer",
            "title": "Indexed Count"
          },
          "unindexed_count": {
            "type": "integer",
            "title": "Unindexed Count"
          },
          "existing_batches": {
            "type": "integer",
            "title": "Existing Batches"
          },
          "proposal": {
            "items": {
              "$ref": "#/components/schemas/ProposedBatch"
            },
            "type": "array",
            "title": "Proposal"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skipped"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "offending": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Offending"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blocker",
          "schemes",
          "scheme",
          "image_count",
          "indexed_count",
          "unindexed_count",
          "existing_batches",
          "proposal",
          "skipped",
          "warnings",
          "error",
          "offending",
          "token"
        ],
        "title": "IndexStatus",
        "description": "``blocker`` explains why indexing can't be shown (paths not configured, input folder missing)."
      },
...
      "OcrStatus": {
        "properties": {
          "blocker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Providers"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "grounding": {
            "type": "boolean",
            "title": "Grounding"
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/BatchOut"
            },
            "type": "array",
            "title": "Batches"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "missing_images": {
            "type": "integer",
            "title": "Missing Images"
          },
          "to_process": {
            "type": "integer",
            "title": "To Process"
          },
          "waiting": {
            "type": "integer",
            "title": "Waiting"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blocker",
          "providers",
          "provider",
          "grounding",
          "batches",
          "total",
          "processed",
          "failed",
          "missing_images",
          "to_process",
          "waiting"
        ],
        "title": "OcrStatus"
      },
      "PageIn": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key"
        ],
        "title": "PageIn"
      },
...
      "RotateIn": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "top_points": {
            "type": "string",
            "enum": [
              "left",
              "right",
              "down"
            ],
            "title": "Top Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "top_points"
        ],
        "title": "RotateIn"
      },
...
      "SaveGroupingOut": {
        "properties": {
          "changed": {
            "type": "boolean",
            "title": "Changed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "changed"
        ],
        "title": "SaveGroupingOut"
      },
...
      "StartOcrIn": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "reprocess": {
            "type": "boolean",
            "title": "Reprocess",
            "default": false
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "StartOcrIn"
      },
      "StartParseIn": {
        "properties": {
          "extractor": {
            "type": "string",
            "title": "Extractor"
          },
          "reprocess": {
            "type": "boolean",
            "title": "Reprocess",
            "default": false
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 0
          },
          "custom_instruction": {
            "type": "string",
            "title": "Custom Instruction",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "extractor"
        ],
        "title": "StartParseIn"
      },
...
    }
  }
}